Add program/faculty/school related info and mock data

+++ user/src/main/java/com/cove/user/controllers/StudentController.java
@@ -41,7 +41,7 @@ public class StudentController {
         return allStudents;
     }
     @RequestMapping("/{id}")
-    public StudentDTO getStudent(@PathVariable final int id){
+    public StudentDTO getStudent(@PathVariable final long id){
         return studentService.getStudentById(id);
     }
